Refactoring All BaseView Fixes #68
