+++ app/src/main/java/com/squarecheck/base/view/BaseView.java
@@ -4,4 +4,10 @@ public interface BaseView<T> {
     void setPresenter(T presenter);
 
     void initView();
+
+    void startLoading();
+
+    void endLoading();
+
+    void showError(String message);
 }
