+++ app/src/main/java/com/squarecheck/lecturer/contract/LecturerAttendanceSummaryDetailContract.java
@@ -10,13 +10,7 @@ import com.squarecheck.student.model.PresenceModel;
 import java.util.List;
 
 public interface LecturerAttendanceSummaryDetailContract {
-    interface View extends BaseView<LecturerAttendanceSummaryDetailContract.Presenter> {
-        void startLoading();
-
-        void endLoading();
-
-        void showError(String errorMessage);
-
+    interface View extends BaseView<Presenter> {
         void showStudentAttendances(List<PresenceModel> data);
 
         void showTitle(Title title);
