+++ app/src/main/java/com/squarecheck/lecturer/contract/LecturerDashboardContract.java
@@ -10,18 +10,14 @@ import java.util.List;
 
 public interface LecturerDashboardContract {
     interface View extends BaseView<Presenter> {
-        void startLoading();
-
-        void endLoading();
-
-        void showError(String errorMessage);
-
         void showSchedules(List<ScheduleModel> schedules);
 
         void redirectToLogin();
 
         void showLogoutConfirmation();
+
         void showDetailProfile(LecturerModel student);
+
         void showProfileImage(String imgPath);
     }
 
@@ -29,7 +25,9 @@ public interface LecturerDashboardContract {
         void requestSchedules();
 
         void logout();
+
         void requestDetail();
+
         void requestProfileImage();
     }
 
@@ -37,7 +35,9 @@ public interface LecturerDashboardContract {
         void requestSchedules(RequestCallback<List<ScheduleModel>> callback);
 
         void logout();
+
         void requestDetail(RequestCallback<LecturerModel> requestCallback);
+
         void requestProfileImage();
     }
 }
