+++ app/src/main/java/com/squarecheck/lecturer/view/LecturerAttendanceSummaryDetailFragment.java
@@ -5,7 +5,6 @@ import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -68,11 +67,6 @@ public class LecturerAttendanceSummaryDetailFragment extends BaseFragment<Lectur
     @Override
     public void endLoading() {
 
-    }
-
-    @Override
-    public void showError(String errorMessage) {
-        Toast.makeText(getContext(), errorMessage, Toast.LENGTH_SHORT).show();
     }
 
     @Override
