+++ app/src/main/java/com/squarecheck/student/view/StudentAttendanceDetailFragment.java
@@ -5,7 +5,6 @@ import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -87,11 +86,6 @@ public class StudentAttendanceDetailFragment
         binding.attendanceRecycler.setVisibility(View.VISIBLE);
         binding.attendanceChart.setVisibility(View.VISIBLE);
         binding.progressCircular.setVisibility(View.GONE);
-    }
-
-    @Override
-    public void showError(String message) {
-        Toast.makeText(getContext(), message, Toast.LENGTH_SHORT).show();
     }
 
     @Override
