+++ app/src/main/java/com/squarecheck/student/view/StudentDashboardFragment.java
@@ -7,7 +7,6 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
-import android.widget.Toast;
 
 import androidx.annotation.Nullable;
 import androidx.recyclerview.widget.LinearLayoutManager;
@@ -78,11 +77,6 @@ public class StudentDashboardFragment extends BaseFragment<StudentDashboardActiv
     @Override
     public void endLoading() {
 
-    }
-
-    @Override
-    public void showError(String errorMessage) {
-        Toast.makeText(getContext(), errorMessage, Toast.LENGTH_SHORT).show();
     }
 
     @Override
